moved chunk visualisation to own module
