+++ crates/map_generation/src/map_generation.rs
@@ -1,6 +1,6 @@
 use bevy::prelude::*;
 use bevy_ecs_tilemap::TilemapPlugin;
-use common::{constants::TILE_SIZE, states::AppState, traits::AddNamedObserver};
+use common::{constants::TILE_SIZE, states::AppState};
 
 #[derive(Default, Reflect, Resource)]
 #[reflect(Resource)]
@@ -9,7 +9,7 @@ pub struct WorldGenerationSettings {
 }
 
 use crate::{
-    ChunkVisualisation, chunk_visualisation,
+    chunk_visualisation,
     messages::{BlockUpdate, UpdateMap},
     world_map::WorldMap,
 };
@@ -17,23 +17,16 @@ use crate::{
 pub fn plugin(app: &mut App) {
     app.register_type::<WorldMap>()
         .register_type::<WorldGenerationSettings>()
-        .register_type::<ChunkVisualisation>()
         .add_message::<UpdateMap>()
         .add_message::<BlockUpdate>()
         .insert_resource(ClearColor(Color::srgb_u8(50, 45, 52)))
         .add_plugins(TilemapPlugin)
+        .add_plugins(chunk_visualisation::plugin)
         .add_systems(OnEnter(AppState::MainGame), spawn_world)
         .add_systems(
             Update,
-            (
-                handle_messages,
-                chunk_visualisation::update,
-                chunk_visualisation::request,
-                chunk_visualisation::delete,
-            )
-                .run_if(in_state(AppState::MainGame)),
-        )
-        .add_named_observer(chunk_visualisation::on_insert, "on_chunk_vis_insert");
+            (handle_messages,).run_if(in_state(AppState::MainGame)),
+        );
 }
 
 fn spawn_world(world_generation_settings: Res<WorldGenerationSettings>, mut commands: Commands) {
